generate section if with Section_ prefix
