+++ src/lib/constants.ts
@@ -9,3 +9,6 @@ export const ERROR_MESSAGE = {
     PROVIDE_ITEM_NAME_OR_ID: "Please provide either the item name or its ID."
 }
 
+export const ID_PREFIX = {
+    SECTION: "Section_",
+};
