+++ src/lib/utils/index.ts
@@ -1,4 +1,3 @@
 export * as QueryBuilder from './query-builder';
 export * from './http-error-factory';
-
-export const isValidId = (id: string): boolean => /^[a-z0-9]{26}$/.test(id);
+export * from "./id";
